GG-29275 WIP

+++ tcbot-engine/src/main/java/org/apache/ignite/tcbot/engine/cleaner/Cleaner.java
@@ -2,10 +2,15 @@ package org.apache.ignite.tcbot.engine.cleaner;
 
 import java.io.File;
 import java.util.Map;
+import java.util.concurrent.Executors;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
 import javax.inject.Inject;
 import org.apache.ignite.ci.teamcity.ignited.buildcondition.BuildConditionDao;
 import org.apache.ignite.ci.teamcity.ignited.change.ChangeDao;
 import org.apache.ignite.ci.teamcity.ignited.fatbuild.FatBuildCompacted;
+import org.apache.ignite.tcbot.common.conf.TcBotWorkDir;
 import org.apache.ignite.tcbot.engine.chain.BuildChainProcessor;
 import org.apache.ignite.tcbot.engine.conf.ITcBotConfig;
 import org.apache.ignite.tcbot.engine.defect.DefectsStorage;
@@ -20,12 +25,14 @@ import org.apache.ignite.tcignited.buildlog.BuildLogCheckResultDao;
 import org.apache.ignite.tcignited.buildref.BuildRefDao;
 import org.apache.ignite.tcignited.history.BuildStartTimeStorage;
 import org.apache.ignite.tcignited.history.SuiteInvocationHistoryDao;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 public class Cleaner {
     @Inject IIssuesStorage issuesStorage;
     @Inject FatBuildDao fatBuildDao;
     @Inject ChangeDao changeDao;
-    @Inject SuiteInvocationHistoryDao suiteHistoryDao;
+    @Inject SuiteInvocationHistoryDao suiteInvocationHistoryDao;
     @Inject BuildLogCheckResultDao buildLogCheckResultDao;
     @Inject BuildRefDao buildRefDao;
     @Inject BuildStartTimeStorage buildStartTimeStorage;
@@ -39,34 +46,85 @@ public class Cleaner {
     @Inject IUserStorage userStorage;
     @Inject ITcBotConfig cfg;
 
+    /** Logger. */
+    private static final Logger logger = LoggerFactory.getLogger(Cleaner.class);
+
+    private final AtomicBoolean init = new AtomicBoolean();
+
+    private ScheduledExecutorService executorService;
+
+    public Cleaner() {
+        suiteInvocationHistoryDao.init();
+        buildLogCheckResultDao.init();
+        buildRefDao.init();
+        buildStartTimeStorage.init();
+        buildConditionDao.init();
+    }
+
     public void clean() {
-        long period = 1;
-        int countEntryToRemove = 1;
-        long oldDate = System.currentTimeMillis() - period;
-        Map<Long, FatBuildCompacted> oldBuilds = fatBuildDao.getOldBuilds(oldDate, countEntryToRemove);
+        try {
+            long period = 1;
+            int countEntryToRemove = 1;
+            long oldDate = System.currentTimeMillis() - period;
+            Map<Long, FatBuildCompacted> oldBuilds = fatBuildDao.getOldBuilds(oldDate, countEntryToRemove);
 
-        for (Map.Entry<Long, FatBuildCompacted> buildEntry : oldBuilds.entrySet()) {
-            long buildCacheKey = buildEntry.getKey();
-            int srvCode = BuildRefDao.cacheKeyToSrvId(buildCacheKey);
-            int buildId = BuildRefDao.cacheKeyToBuildId(buildCacheKey);
-            String strSrvCode = compactor.getStringFromId(srvCode);
+            for (Map.Entry<Long, FatBuildCompacted> buildEntry : oldBuilds.entrySet()) {
+                long buildCacheKey = buildEntry.getKey();
+                int srvCode = BuildRefDao.cacheKeyToSrvId(buildCacheKey);
+                int buildId = BuildRefDao.cacheKeyToBuildId(buildCacheKey);
+                String strSrvCode = compactor.getStringFromId(srvCode);
 
-            suiteHistoryDao.remove(buildCacheKey);
-            buildLogCheckResultDao.remove(buildCacheKey);
-            buildRefDao.remove(buildCacheKey);
-            buildStartTimeStorage.remove(buildCacheKey);
-            buildConditionDao.remove(buildCacheKey);
+                suiteInvocationHistoryDao.remove(buildCacheKey);
+                buildLogCheckResultDao.remove(buildCacheKey);
+                buildRefDao.remove(buildCacheKey);
+                buildStartTimeStorage.remove(buildCacheKey);
+                buildConditionDao.remove(buildCacheKey);
 
-            defectsStorage.removeOldDefects(oldDate, countEntryToRemove);
-            issuesStorage.removeOldIssues(buildId, strSrvCode, countEntryToRemove);
+                defectsStorage.removeOldDefects(oldDate, countEntryToRemove);
+                issuesStorage.removeOldIssues(oldDate, countEntryToRemove);
 
-            File logsDirectory = new File(cfg.getTeamcityConfig(strSrvCode).logsDirectory());
-            File[] buildLogFiles = logsDirectory.listFiles((dir, name) -> name.contains(Integer.toString(buildId)));
-            for (File file : buildLogFiles) {
-                file.delete();
+                final File workDir = TcBotWorkDir.resolveWorkDir();
+
+                for (String srvId : cfg.getServerIds()) {
+                    File srvIdLogDir = new File(workDir, cfg.getTeamcityConfig(srvId).logsDirectory());
+                    for (File file : srvIdLogDir.listFiles()) {
+                        if (file.lastModified() < oldDate)
+                            file.delete();
+                    }
+                }
+
+//            File logsDirectory = new File(cfg.getTeamcityConfig(strSrvCode).logsDirectory());
+//            for (File file : logsDirectory.listFiles()) {
+//                if (file.lastModified() < oldDate)
+//                    file.delete();
+//            }
+
+                fatBuildDao.remove(buildCacheKey);
             }
+        }
+        catch (Exception e) {
+            e.printStackTrace();
 
-            fatBuildDao.remove(buildCacheKey);
+            logger.error("Failure periodic check failed: " + e.getMessage(), e);
+        }
+    }
+
+    public void startBackgroundClean() {
+        try {
+            if (init.compareAndSet(false, true)) {
+
+                executorService = Executors.newScheduledThreadPool(1);
+
+                executorService.scheduleAtFixedRate(this::clean, 0, 10, TimeUnit.SECONDS);
+
+            }
+        }
+        catch (Exception e) {
+            e.printStackTrace();
+
+            init.set(false);
+
+            throw e;
         }
     }
 }
